Refactor: footer - step 관련 코드들 커스텀 훅으로 분리
